Added current available data with naming...
convention: "out_n_k.txt" where "n" is the zeros parameter, n:n+100 corresponding to the "enumerate" line in H_t_probable... and "k" is the t parameter running through t=k:.50 with a .01 step. Also included is an edited H_t_probable file which serves as a sample of the input in each case.