upload/update data (computed with prev w/o complex lib)
out_300_10 and out_400_15 are calculated with t=.10 to .46 and t=.15 to .43 resp. out_300_20 is from t=.20 to .49. Note all of these files have an artifact from one segment of the code that may have been commented out, namely the first few lines in the data that are not in a table.